Hide Duo availability for Duo Core only scenarios
When the customer has only Duo Core, we want to hide the Duo
availability options. We only want to display it for Duo Pro and Duo
Enterprise. Additionally we want to change the title depending on the
existing Duo Pro or Duo Enterprise add-on.

Subgroups don't have Duo add-ons, but we still want to display the Duo
availability options for them. To solve this we can iterate the
group's parents to find the root ancestor.

Changelog: added
EE: true