Drop index (expire_at, job_id) from p_ci_job_artifacts
This migration synchronously drops the index from
p_ci_job_artifacts which matches the definition
(expire_at, job_id). This index may have a different name
on different instances, so we cannot rely on dropping it by
name.

This index is unused so we are dropping it to recover
disk space.

This index was dropped asynchronously for GitLab.com

Changelog: other