Fix ConcurrencyLimitSampler lease exclusiveness
Currently, we're checking on lease existence whether a process holding
the lease should continue writing the metrics with
ExclusiveLease#exists?.

However, when the lease expires another process can obtain the lease and
extend the lease for another 300s. The lease having the same key, makes
the earlier process seem to still holding the lease, thus
continuing to write the metrics forever.

This is because ExclusiveLease#exists? only checks whether the lease
with the same key exists, without comparing the uuid.

This commit changes it to use ExclusiveLease#same_uuid? from EE
implementation that checks whether we still have the same uuid
of the lease.

Also moved the #same_uuid? implementation to FOSS ExclusiveLease.