Set FF allow_duo_base_access to true
We plan set this FF in the future to true. Also it will be removed at
some point and with that Self-Managed instances using the release 18.0
should have the default value set to true and with that being able to
use the feature, even if the FF has been deleted from any remote source.

EE: true
Changelog: changed