Add strict validation of error responses against starboard API specs
Recently, we had a report of an issue where the GitLab Agent's starboard
client could not create vulnerabilities. It received a 400 response for
specific reports, but the logging showed nothing about the response
body. We want the client to include the specific error in the logging,
so we attempted to re-use the error response bodies that would
presumably be found in the API specs. Unfortunately, there were no
response bodies to be found, and no verification that the errors matched
our expectations.

This commit changes that so that we now **strictly** check against the
response body, and if that fails, we print out the expected response
body, so that we can make a decision on whether to update the
expectation, or update the endpoint/service.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/536190