Remove specs related to licenses with an empty start date
Prior to this change, it was discovered that a validation during the
license creation prevents licenses from having no start date. Any specs
covering this use case on the GL side would cover an invalid license and
can be removed.

This change removes any specs related to a created license with no start
date since it's not a valid scenario.