Expose namespace link paths on GraphQL
This will enable the ~frontend to manage and handle better resources on
the frontend. The initial goal is to give better support for the
WorkItems UI, but this can be extended to other features on GitLab.

To make it extensible, the `Types::Namespaces::LinkPaths` is an
interface, this way the values are implemented polymorphically on:
- `Types::Namespaces::UserNamespaceLinksType`,
  - `Types::Namespaces::ProjectNamespaceLinksType`
  - `Types::Namespaces::GroupNamespaceLinksType`

  Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/537438
  Changelog: added