Put Duo Vulnerability Resolution patching logic in a service
Context
-------------------

We have quite a few customer contacts about the Duo Vulnerability
feature failing to create and MR. We also have a bunch of these errors
in the logs that appear to be related.

After an investigation[^1] into the root cause, we realized that
[this line][1], which is used to `gsub` the llm fix into the source
file, is too brittle.

We have examples in the logs where we would expect a patch to be
created, but instead the `gsub` fails to update the source
code (because it fails to detect the old code)

A fix was [drafted][[4] and verified to resolve the issues in staging,
however the risk of changing the patch generation code was decided to
be too high when compared to the amount of this error we see.

This Change
----------------------

The major difference with this change is we are now using
`Files::MultiService`, to modify the source file directly. [^3]

This removes a lot of the hand-rolled patch creation.

In that draft I also used a different find/replace strategy for
applying the llm patch to the source code; moved the find/replace
logic into a service; and added additional error handling and logging

This change extracts the error handling, logging, and new service into
a separate MR. These changes are structural prep work and very low
risk. Moving them to their own change make the bug fix MR more focued
and easier to review

[0]:https://gitlab.com/gitlab-org/gitlab/-/issues/497193#note_2414009651
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/479398d0fafd312a3168bbc8684b1fdbe92d188e/ee/app/services/merge_requests/create_from_vulnerability_data_service.rb#L162
[2]:https://gitlab.com/gitlab-org/gitlab/-/issues/526882#note_2416248934
[3]:https://docs.gitlab.com/api/commits/#create-a-commit-with-multiple-files-and-actions
[4]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/185846

[^1]: The investigation started in [this issue][0] and continued in
      [this issue][2], where the fix was determined
[^2]: "never" is a bit too strong for an LLM response. We haven't seen
      it in production yet and regardless if it happens we would still be
      dealing with the same problem.
[^3]: `Files::MultiService` is the service backing [this API][3]

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/497193
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/526882
Changelog: changed
EE: true