Remove derisk FF: `resolve_vulnerability_send_finding_description`
Context
------------------

We had stopped sending the description of a vulnerability finding when
this feature was first launched, in order to save on tokens.

With newer models, this token constraint is no longer a worry, so we
want to start sending the finding description in the LLM prompt again.

This change
------------------

We added a derisk FF just to be really really sure the description
wouldn't result in LLM errors.

The flag has been globally enabled, and we see no error rate
difference in the [dashboard][0].

[0]:https://log.gprd.gitlab.net/app/r/s/n0eRe

---

closes: https://gitlab.com/gitlab-org/gitlab/-/work_items/537615
EE: true
Changelog: removed