Use timezone related dates in expiring subscription message code
Prior to this change, the code related to the expiring subscription
message was using `Date.today`. This doesn't account for timezones which
can lead to unexpected behavior of the message.

This change replaces all `Date.today` usages in the
`Gitlab::ExpiringSubscriptionMessage` class with `Date.current` to
resolve unexpected behavior related to timezones.

Changelog: changed
EE: true