Focus on username or password field in sign-in tabs
When there are tabs in the sign-in page or admin mode, there is a
mechanism that memoize which tabs was used the last time. However, the
mechanism makes the password or useranme fields lose focus.

This change return the focus to the password field or the username
field, if one available.

Related to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/178285

Changelog: fixed