Log a debug message if a suspected dead code path is taken
Context
------------

I am almost certain this `else` branch, and all of the code under it,
is dead code. I think this because:

1. Our [prompt][0] says to always include `START_OLD_CODE`[^1]
2. The [specs][1] that cover the `if` part of this conditional are in a
   context called `context 'with new response structure'`

I think that this was a migration path for deprecating an old version
of the prompt. As there is no permanent team working on this feature,
I think the deprecation was paused.

I would like to delete the `else` branch, and all the spec code
covering it, to make a bug fix diff have a smaller, more focused diff.

This change
------------

This will be done in a series of 3 MRs. They have all been drafted and
set up to point at each other, so that reviewers can get the full
changeset context:

1. **This MR**
    - this logging will sit in production for a milestone. If no logs
    are recorded, we will move forward with the next MR in the series
2. [MR removing the dead code][3]
3. [A bugfix MR][4]. This MR has a much cleaner diff if all the dead code
   in 2 gets removed

[^1]: Yes, the LLM could still mess up however, in that situation, we
      should log that as a "low quality LLM response" event, and treat this
      as an error.
[^2]: It isn't urgent, so I just want to find out if this code is ever
      actually run in production.

[0]:https://gitlab.com/gitlab-org/security/modelops/applied-ml/code-suggestions/ai-assist/-/blob/f3ea42027c1dde9a8721b0308ff85d5001b845e8/ai_gateway/prompts/definitions/resolve_vulnerability/user/1.0.0.jinja#L40
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/358e12fa4b4fa29247c30abe20633a77f7cc2948/ee/spec/services/merge_requests/create_from_vulnerability_data_service_spec.rb#L471
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/191381
[3]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/191382
[4]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/191349

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/191381
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/536007
Changelog: changed
EE: true