Explain the need to finalize BBM in required stop + 1 milestone
We should finalize the BBMs in the next milestone after the required to
ensure that we give customers upgrading only between the required stops
an opportunity to run the batched background migrations in the
background. Otherwise, finalizing migrations in 17.8 would
force synchronous execution during upgrade from 17.5, causing downtime
if the migrations runtime exceed the maintenance window.