Fix setting confidential and set_parent
This MR resolves an issue where the `/confidential` 
and `/set_parent` quick action params are used in 
conjunction on a work item. 

Previously, the `/set_parent` check for 
confidentiality would not consider the case where 
the `/confidential` quick action was also passed in 
the same template. 

In this case, since the child would also become
confidential, we should allow it to be assigned to a
confidential parent.

Changelog: fixed