Reorg schema of tags table
The tags table is for the CI runners service to function. The CI runners
service is cell-local, so use the correct schema `gitlab_ci_cell_local`.

See https://gitlab.com/gitlab-org/gitlab/-/issues/470872

Some allowlists in the no_cross_db_foreign_keys_spec was tweaked because
of this schema change.