doc: Reinforce that leaked secrets cannot be easily redacted
The documentation for "Prune unreachable objects" currently cite
removing leaked secrets as an example use-case. Because GitLab uses
keep-around refs extensively, commits associated with an MR will not be
pruned with this feature.

Reinforce that when a secret is committed and pushed, there is no way to
guarantee that it can be removed and be "unleaked".