Add YML file of potentially unused methods
This allows us to more transparently track the methods we believe are
unused, which allows for more people to contribute to cleaning them up.
This also allows us to remove the hardcoded `UNUSED_METHODS` constant,
enabling us to generate better error messages with the eventual goal of
adding this script to our automated lefthook checks to help keep our
code from getting so crufty with unused methods in the first place.