Rename OrganizationUserAlias to OrganizationUserDetail
As we plan out the architecture for Organizations and Cells, we have
broadened the concept of per-organization usernames to potentially
include other types of data.

We want to replace the `organization_user_aliases` table with the
`organization_user_details` table. This MR is the first part of that
work, adding the `organization_user_details` table and related
ActiveRecord model. There are some improvements over the previous table,
including using `TEXT` type rather than `VARCHAR` , and adding a `lower`
composite index similar to the `users` table.

Changelog: changed