Zoekt: Only enable search when nodes are online
The `search_enabled?` method in Replica now joins with zoekt nodes and
checks for online status. Previously, it would return true for replicas
with offline nodes, which could lead to unsuccessful searches.

This change ensures that search is only considered enabled if there are
online nodes available to service the search requests.

Changelog: fixed
EE: true