Log when `.safe_find_or_create_by` cannot find a row that should exist
In https://gitlab.com/gitlab-org/gitlab/-/issues/543351 we saw that
the runner API quietly returned a 404 when a replicated table was out
of sync with the source table. Log an exception so that we can see
the full database error for debugging.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/543351

Changelog: changed