Remove duoSettings object from Duo setting update GraphQL mutation
Prior to this change, the GraphQL mutation response to update the Duo
settings was modified to return an `Ai::Setting` object instead of its
individual attributes. Since the fields for the individual attributes
nor the field for the `Ai::Setting` object are consumed by the frontend
yet, the frontend mutation response can be simplified to only return
the errors of the mutation.

This change updates the response of the Duo setting update GraphQL
mutation in the frontend to only return the `errors` by removing the
`duoSettings` field. If the info of the Duo settings is needed in the
future, it can easily be added back into the response. The backend
mutation logic will be kept as is.

Changelog: changed
EE: true