Fix date-dependent DORA specs
Previously the DORA specs failed because the frontend submitted
`start_date` and `end_date` using relative times, but the seeded data
relied on fixed dates because of `time_travel_to`.

We can't mix relative and absolute timestamps. To fix these specs,
avoid the absolute timestamps and add some buffer for the oldest DORA
metrics so that the data doesn't get included in the calculations.

Relates to
https://gitlab.com/gitlab-org/quality/engineering-productivity/master-broken-incidents/-/issues/12590