Fix UX bug related to personal access token scopes
In the user setting personal access token page, a scope is mandatory.
Previously, it was possible to skip the frontend validation if the user
selected a scope and then deselected. That resulted in a backend error.

Now, the frontend correctly detects if an scope has been chosen, under
all possible interactions.

Changelog: fixed