Ignore `require-selections` violations in fragments
`@graphql-eslint/eslint-plugin@4.0.0` now detects `require-selections`
violations within fragment definitions, which uncovered a few offences
in existing fragments of ours. For the sake of efficiency, we are
ignoring those violations for now so that we can proceed with the
package upgrade without changing production code and having to verify
the impact. Comments have been added to explain our reasoning and to
encourage folks to address the violations if they happen to touch the
affected files.