Protected packages: Original package publisher for NuGet worker (Part 1)
In order make the NuGet package protection mechanism more robust,
it was discussed and decided to pass
the original package publisher (user or deploy token) to
the Packages::Nuget::ExtractionWorker.
This allows to correctly apply package protection rules
when the package is pushed by a user or a deploy token.

This was discussed as part of a previous MR review: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/184059#note_2426761343.

Changing the arguments of a worker requires multi-milestone deploytment.
This commit represents the first step of the multi-step deployement,
see https://docs.gitlab.com/development/sidekiq/compatibility_across_updates/#multi-step-deployment.

Changelog: other