Reduces the number of GraphQL queries sent on merge request dashboard
Currently we send mutiple requests to get the tab and list counts even
though they are the same query. This chnages that by sending only 1
request with all the count queries for the tab count and then for the list
counts we only listen to cache events instead of sending the request.
With this change we take the total count of GraphQL queries from 10 to 8.