Updated mobile styling of merge request header
Updates and tidies up some of the styling of the merge request header
for mobile.
Removes a lot of extra classes to do with spacing by instead using flexbox
and gaps to create the spacing we want.
Removes an unnecessary `.to_json` call on the merge request object by only
selecting the items that are required.