Use exclusions in CODEOWNERS
We previously used a quirk in the logic of CODEOWNERS to prevent
some of the files from requiring approval due to the `*` entry on line 5.

CODEOWNERS currently sets any approval rules without any approvers
to auto-approved. 

We recently implemented an exclusion feature which would achieve the same result without needing to populate an approval rule and auto-approve it.