[GraphQL] Only return available WorkItems::Type for the given namespace
To enable frontend to build a UI for users to create WorkItems from a
centralized place, we need to filter the `WorkItem::Type` (WIT) to only
return the available WIT for the given namespace. The filtering logic
is taking in consideration the namespace type (Group/Project) and
feature flags.

Example of GraphQL query:

```graphql
query getNamespaceWorkItemTypes {
  namespace(fullPath: "group/p1") {
    name
    workItemTypes {
      nodes {
        name
      }
    }
  }
  group(fullPath: "group") {
    name
    workItemTypes {
      nodes {
        name
      }
    }
  }
  project(fullPath: "group/p1") {
    name
    workItemTypes {
      nodes {
        name
      }
    }
  }
}
```

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/524828
Changelog: changed