Bundle-URI: Updating public doc
The bundle URI received an upgrade in the recent version of
GitLab, with the auto-generation of bundles. This new feature
is also behind a feature flag. The current documentation
do not expose this new feature and feature flag.

This commit updates the documentation to introduce the
auto-generation of bundles and expose the feature flag
to use. It also removes the `experimemental` status
of the whole bundle URI feature since this feature is
now fully available in the self-served offering.

It will soon be available on GitLab.com, but it is not yet.

References:
https://gitlab.com/gitlab-org/gitaly/-/issues/6709