Protect pipelines of MRs between protectd branches
We have two related open issues caused by the same
root cause that needs to be fixed. The values for
`CI_COMMIT_REF_PROTECTED predefined CI variable` and `value for
ref_protected claim in CI JWT ID Token` are wrongly set as `false` in
one scenario when it is supposed to be true - i.e. in the scenario where
a pipeline is run for an Merge Request whose source and
target branches are protected. This commit fixes the issue.

We are setting the pipeline of a Merge Request to
be protected if both the source and target branches are protected.
To make sure existing pipelines are not affected, we are only enabling
this change based on a project setting. This setting will be enabled by
default for new projects. By default this setting is disabled
for existing projects.

Documentation changes will be made in a follow up
MR after the rollout of the feature flag.

The changes are behind a derisk feature feature flag

Issue - https://gitlab.com/groups/gitlab-org/-/epics/17186

Changelog: changed