Fix Internal Server error from branchRuleCreate mutation
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/508212

**Problem**

The presence check for project is missing for branchRuleCreate mutation.
Also, the permission to create a branch rule is not enforced on mutation
level.

**Solution**

Use `authorized_find!` to verify the presence of provided project path
and user permissions to create a branch rule.

Changelog: fixed