Present queue_size and queue_depth in CI job payload
As described in
https://gitlab.com/gitlab-org/ci-cd/shared-runners/infrastructure/-/issues/298,
current metrics tracking the size and depth of the potential queue of CI
jobs for a given runner is very limited. The two main problems are:

1. It's a generalized number for the whole GitLab instance, partitioned
   by three runner types and nothing else.

2. It's a metric exported by GitLab Rails internal exporter, which means
   in most cases it's not available for the users.

The second problem is especially important, as unless you're a GitLab
instance owner, you have no visibility of the queue (even very
generalized one) that potentially aims your self-hosted runners.

Some time ago we've [updated the API](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/90653)
to pass information about queuing duration to Runner, inside the job
payload.

Here I'm proposing a follow-up for that. We're going to present the
`queue_size` and `queue_depth` values, being exactly the same thing that
we're counting on the GitLab side. Here, however, that information is
made available to GitLab Runner.

In a second MR I'll introduce a change in GitLab Runner, where it
consumes these values and exports them as dedicated metrics, that any
runner owner can track.

Changelog: changed