Add condition to skip stdout silencing based on env variable
Problem: When the example metadata `:silence_stdout` is used
then the standard output is redirected to reduce console output noise.
But, it can be problematic when you want to debug tests and
implementaion code with the `debugger` command.
In this cases, interactive debugger console does not behave correctly
because the standard output is redirected, too.

Solution: Add an environment variable `SKIP_SILENCE_STDOUT` that allows
skipping the output silencing.

Changelog: other