Update Groups API status parameters to respect inherited status
Previously, the status filters (active, archived, aimed_for_deletion)
only checks for the group's status. It does not check for the status
of its parents and ancestors.

This update ensures that the filter is also checks the group's ancestor
status. If a group is inactive, all of its subgroups should also be
recognized as inactive.

Changelog: changed