Do not make membership of same level as inherited
This reverts commit 3503a17731a0fe980268db04e13cb73f2f01fbf8.

In Direct Transfer, as we process a parent group, we query for all
members of the group through GraphQL. As we encounter the member in
`MembersPipeline`, we create placeholder membership records.

We then do the same for projects in the group.

Where a user has a direct membership in the parent group, and therefore
and inherited membership of the project, this leads to inherited
memberships being encountered twice:

- Once when processing the parent group (the direct membership)
- A second time when processing the project (the inherited membership)

When user mapping is disabled, the `MembersPipeline` will check that any
inherited membership levels are < to the one being imported before
creating a new (direct) one
https://gitlab.com/gitlab-org/gitlab/-/blob/331dad5654cc4f1a51f225fc05be8c717096a401/lib/bulk_imports/common/pipelines/members_pipeline.rb#L55-56

But when user mapping is enabled, we create a (direct) membership if the
existing inherited levels are <= an inherited one.
This was an intentional implementation in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/167396.

This results in us creating a direct membership in the parent group, and
then a direct membership in the project, always.

This change reverts https://gitlab.com/gitlab-org/gitlab/-/merge_requests/167396
to only create a direct membership when less than an existing inherited
membership.

Changelog: fixed