Ensure parent_link on epic creation
This adds a validation for the parent_link FK on epics to ensure
we set it when saving the model.

We already set parent_link when syncing it from
WorkItems::ParentLink::CreateService. Before we introduce a DB
constraint and backfill the data, we're ensuring we're setting it
every time on the Model.