Avoid Gitaly race conditions in Workhorse integration tests
The previous Gitaly integration tests always used `foo/bar.git` as the
repository path and removed it at the start of the test.  However,
it's possible that some background `git` command or Goroutine might
run and alter the state of the repository and interfere with
the state, leading to intermittent test failures.

To avoid this, let's just use a random path for every test
and ensure that we clean up the repository after each test.

Relates to
https://gitlab.com/gitlab-org/quality/engineering-productivity/master-broken-incidents/-/issues/13363