Use inner join to avoid attempting to save incomplete records
Some self-managed instances reported that this migration was failing each batch
due to a not null constraint on `namespace_id` being disrespected.

This is the result of the LEFT JOIN on vulnerability_historical_statistics
sometimes not having an appropriate project to match against. By using an INNER
JOIN we ensure that only full records are attempted to be updated