Document registry public schema issues
If a user encounters issues with the public schema on Postgres 15+,
they should ensure that the database username is an owner of the
database that it will manage.

This documentation change specifies how to make the registry
user an owner of the logical database.

Related to https://gitlab.com/gitlab-org/container-registry/-/issues/1497.