Rescue NoteDiffFileCreation error in worker
The `CreateNoteDiffFileWorker` is being enqueued whenever
`DiffNote#diff_file` is called and it has no `note_diff_file` yet.
But we are already creating the `note_diff_file` on create.

The `NoteDiffFileCreationError` error is being raised when we can't
find the diff file or diff line (based on `original_position` of the
`DiffNote`). If we can't find it on create, then we won't be able to
find it even when calling `DiffNote#diff_file` again.

With that said, retrying the `CreateNoteDiffFileWorker` multiple times
won't make any difference. We log and track the error so it won't
trigger a Sidekiq retry and the job won't fail.