Implement EligibleApproverGroupsFilter logic
This change implements logic to find eligible approver groups for code
owner entries.

A group is eligible when it contains at least one user who can approve
an MR for the given project.

To ensure this check is performant we preload everything required to
perform these checks.

In a follow-up we will add this filter to the list of filters that are
processed when validating code owners.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/521152+