Change retries for pipeline processing to 25
If there is any reason processing fails with an error
we have more chance of the application and it's data recovering
and healing itself if we keep trying over multiple days.

For example, if an incident led to database/redis errors that were
temporary such as timeouts or out of memory, we would retry with exponential
backoff off which would allow it to retry the next day once the incident is resolved.

Since the worker is idempotent and the worker checks that the jobs/pipeline
need_processing we can safely executing this a second or even 25th time
in the case of an intermitent error.