Downgrade Workhorse PubSub EOF errors to info messages
Previously if the Redis server shut down an idle Redis PubSub
channel, Workhorse would log an error message:

```
{"error":"keywatcher: pubsub receive: EOF","level":"error","msg":"",...}
```

This message gives the impression there may be a problem, but in
reality this message is expected if:

1. Workhorse has not received an `/api/v4/jobs/request` from a runner
within the Redis TIMEOUT, which defaults to 60 seconds in Omnibus
GitLab packages.

2. Workhorse is not handling long polling requests.

With Cloud Native GitLab pods, runners often keep a long-lived HTTP/2
connection to a single Workhorse, so other Workhorse pods might not
handle any job requests for 60 seconds.

The redis-go internal logger was already logging this EOF error to
stderr. This commit directs all redis-go log messages to the Workhorse
info output.

Changelog: changed