Fix: Ensure placeholder users are organization users
While verifying data related to Organization Membership, it
turned out that since the deploy of this[1] MR, some users do not
have an organization associated with them.

The logic in Gitlab::Import::PlaceholderUserCreator was using User
model directly. And the referenced MR removed the logic for
organization association.

We can address this by using Users::AuthorizedBuildService for
creating the placeholder users. This class will also take care of
namespace creation

[1] https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175288

Changelog: fixed