Finalize BBM CreateMissingTerraformModuleMetadata
This migration was finished at `2024-11-22 06:56:02 UTC`, you can confirm
the status using our
[batched background migration chatops commands](https://docs.gitlab.com/ee/development/database/batched_background_migrations.html#monitor-the-progress-and-status-of-a-batched-background-migration).
  To confirm it is finished you can run:

```
/chatops run batched_background_migrations status 2000752 --database main
```

The last time this background migration was triggered was in
[db/post_migrate/20241118145604_queue_create_missing_terraform_module_metadata.rb](https://gitlab.com/gitlab-org/gitlab/-/blob/master/db/post_migrate/20241118145604_queue_create_missing_terraform_module_metadata.rb)

You can read more about the process for finalizing batched background migrations in
https://docs.gitlab.com/ee/development/database/batched_background_migrations.html .

As part of our process we want to ensure all batched background migrations
have had at least one
[required stop](https://docs.gitlab.com/ee/development/database/required_stops.html)
to process the migration. Therefore we can finalize any batched background migration that was added before the
last required stop.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::OverdueFinalizeBackgroundMigration keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: other