test: Refactor factory usage of project_deploy_token
While working on the package registry tests, I noticed that
the `project_deploy_token` factory was not necessary
for most of the tests.
By adjusting the usage of the factory `deploy_token`,
it is possible to avoid the usage of the `project_deploy_token` factory
in most test cases.

Changelog: other