Ensure file tree browser renders correctly without refresh
In this MR, we add logic to ensure file tree browser renders correctly
without refresh when navigating between overview, blob, tree.

Fetching and render files in the tree view was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/190487.
This was done by initiating the FTB app in _blob.html.haml.
However, when navigating from the repository files list(_files.html.haml)
to a blob we only swap out the content area in order to avoid a full
reload. This MR adds logic to initialize FTB there.