Limit the number of pipeline count in merge request table
Previously GitLab would count the number of CI pipelines
associated with a ref when creating a merge request. This SQL
query could time out and prevent a merge request from being created.

To avoid this database timeout, limit the number of CI pipeline count
to 100. If there are more than 100 CI pipelines, the tab will show
"100+".

Relates to
https://gitlab.com/gitlab-com/request-for-help/-/issues/2928

Changelog: fixed