Add workhorse circuit breaker feature flag and config field
Since feature flags are controlled by rails, we can return the flags
value in a response for workhorse to use. The feature flag can toggle
whether the circuit breaker is able to trip or not.

Since the feature flag cannot control errors that occur before a request
to rails, I've  also included an EnableCircuitBreaker config value that
will be used to toggle the entire circuit breaker on or off.

Also fixes a bug where the delegate roundTrip body was prematurely
closed.