Fix Protected Tags show page
There's currently a bug with the protected tags show page as it returns
a 404. This happens because we're missing a show page for the protected
tags.

**Solution**

Add the missing page so we can render successfully. Make the quarantine
on the protected_tags spec more specific to the flaky specs to
re-introduce spec coverage.

Changelog: fixed