Add a `jsonb` column to `vulnerability_exports` table
Context
------------------

We are building a PDF reporting feature in tandem with the new
vulnerability dashboard.

We are adding many different data visualizations to the web. You can
adjust the data with various filters, such as time range.

We want to show those same visualization, with the same filter states,
in the PDF.

To do this without having two maintain 2 separate UI component
libraries (web and PDF), and keeping them in sync, we are going to
send the SVGs in the body of the PDF export request, and render them
in the PDF.

This change
------------------

During a [spike][0], where we tested out the feasibility of this
approach, we realized we would need to store the SVG data until the
PDF export worker is running.[^1]

Some notes about the exports:

- exports are scheduled for deletion after the file is generated, so
  this change should not result in too much DB pressure
- This jsonb data does not need to be queried or indexed

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/192439#note_2527778255

[^1]: You can see a recording of demo of using the SVGs to create the
      PDF report in [this thread][0] on the spike issue.

---

epic: https://gitlab.com/groups/gitlab-org/-/epics/16989
resolve: https://gitlab.com/gitlab-org/gitlab/-/issues/546360
Changelog: added
EE: true