Fix issue with non-supported dependency packager
This commit addresses an bug within our
GraphQL API.

Currently we only support a limited amount
of dependency packagers (ENUM). If a project
uses a packager that is not part of this list,
the GraphQL-API returns a type error (500).

This change filters out non-supported values
and returns `nil` instead.

Changelog: changed
EE: true