Update Projects API status parameters to respect inherited status
Previously, the status filters (active, archived, aimed_for_deletion)
only checks for the project's direct status.
It does not check for the status
of its ancestor groups.

This update ensures that the filter is also checked against the
project's entire ancestry chain. If a group is inactive, all of its
projects should also be
recognized as inactive.

Changelog: changed