fix: Subtle test bug with counter assigned to a constant
Constants are not recycled between test runs.

When metrics are assigned like this:

```ruby
OMNIAUTH_LOGIN_TOTAL_COUNTER =
          Gitlab::Metrics.counter(:gitlab_omniauth_login_total, 'Counter of initiated OmniAuth login attempts')
```

This won't be problematic in production, but in tests it does, because
the registry can be reset between test runs, but constants are never
recycled. This constant will keep the reference to the previous
registry.