bug: Fix mermaid rendering on wiki notes
Problem: At the moment, mermaid diagrams are not rendered correctly
in wiki comments.
This is caused because the `renderGFM` function is not called
when the note is updated.
See https://gitlab.com/gitlab-org/gitlab/-/issues/547843.

Solution: This commit adds a watcher on the `note` property
to call `renderGFM` whenever the note is updated.

Changelog: fixed