Cleanup FF allow_issuer_claim_custom....id_tokens
We had safely released the feature that allows
self hosted instances to customize the issuer claim
for CI ID tokens. The feature was put behind a derisk flag
to allow us to safely roll it out on .com. Now that we have
safely rolled it out, we can remove the derisk flag and
cleanup the related code. This commit does that.

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/544845)

Changelog: changed